--- kubectl-sql.json
@@ -1,72 +1,0 @@
-{
-    "aliases": {
-        "owner.name": "metadata.ownerReferences[1].name",
-        "owner.kind": "metadata.ownerReferences[1].kind",
-        "hostIP": "status.hostIP",
-        "podIP": "status.podIP"
-    },
-    "table-fields": {
-        "VirtualMachineInstance": [
-            {
-                "title": "NAMESPACE",
-                "name":  "namespace"
-            },
-            {
-                "title": "NAME",
-                "name":  "name"
-            },
-            {
-                "title": "PHASE",
-                "name":  "status.phase"
-            },
-            {
-                "title": "IP",
-                "name":  "status.interfaces[1].ipAddress"
-            },
-            {
-                "title": "CREATION_TIME(RFC3339)",
-                "name":  "created"
-            }
-        ],
-        "VirtualMachineInstanceMigration": [
-            {
-                "title": "NAMESPACE",
-                "name":  "namespace"
-            },
-            {
-                "title": "NAME",
-                "name":  "name"
-            },
-            {
-                "title": "VMI",
-                "name":  "spec.vmiName"
-            },
-            {
-                "title": "CREATION_TIME(RFC3339)",
-                "name":  "created"
-            }
-        ],
-        "ReplicaSet": [
-            {
-                "title": "NAMESPACE",
-                "name":  "namespace"
-            },
-            {
-                "title": "NAME",
-                "name":  "name"
-            },
-            {
-                "title": "READY",
-                "name":  "status.readyReplicas"
-            },
-            {
-                "title": "REPLI",
-                "name":  "status.replicas"
-            },
-            {
-                "title": "CREATION_TIME(RFC3339)",
-                "name":  "created"
-            }
-        ]
-    }
-}
