+++ pkg/cmd/constants.go
@@ -27,7 +27,7 @@ var (
 	clientVersion = "GIT-master"
 
 	// sql command.
-	sqlCmdLong = `Uses SQL-like language to filter and display one or many resources.
+	sqlCmdLong = `Uses SQL-like language to filter and display Kubernetes resources.
 
   kubectl sql prints information about kubernetes resources filtered using SQL-like query`
 
@@ -39,33 +39,9 @@ var (
   
   # List first 5 pods ordered by creation time in descending order (newest first).
   kubectl sql "select * from pods order by created desc limit 5"
-  
-  # Print current available aliases.
-  kubectl sql aliases
 
   # Print this help message.
   kubectl sql help`
-
-	// sql get command.
-	sqlGetShort = "Uses SQL-like language to filter and display one or many resources"
-	sqlGetLong  = `Uses SQL-like language to filter and display one or many resources.
-
-  kubectl sql get prints information about kubernetes resources filtered using SQL-like query.`
-
-	sqlGetExample = `  # List all pods in table output format.
-  kubectl sql get pods
-	
-  # List all replication controllers and services in json output format.
-  kubectl sql get rc,services --output json
-
-  # List all pods where name starts with "test-" case insensitive.
-  kubectl sql select * from pods where name ilike 'test-%%'"
-
-  # List all pods where the memory request for the first container is lower or equal to 200Mi.
-  kubectl sql "select * from */pods where spec.containers[1].resources.requests.memory <= 200Mi"
-  
-  # List pods ordered by name, limiting to 10 results.
-  kubectl sql "select * from pods order by name limit 10"`
 
 	// Errors.
 	errUsageTemplate = "bad command or command usage, %s"
