--- sql.yaml
@@ -1,27 +1,0 @@
-apiVersion: krew.googlecontainertools.github.com/v1alpha2
-kind: Plugin
-metadata:
-  name: sql
-spec:
-  version: v0.3.21
-  homepage: https://github.com/yaacov/kubectl-sql
-  platforms:
-  - selector:
-      matchLabels:
-        os: linux
-        arch: amd64
-    uri: https://github.com/yaacov/kubectl-sql/releases/download/v0.3.21/kubectl-sql.tar.gz
-    sha256: d73d661558b49561912525068da724ef3ab6cd5f1925a6df5300bc6beaa18df6
-    files:
-    - from: "*"
-      to: "."
-    bin: kubectl-sql
-  shortDescription: Use SQL like language to query the Kubernetes cluster manager.
-  description: |
-    This plugin use SQL like language to query the Kubernetes cluster manager.
-  caveats: |
-    Usage:
-      $ kubectl sql
-    For additional options:
-      $ kubectl sql --help
-      or https://github.com/yaacov/kubectl-sql
